requirements.txt
anyio
astor
beautifulsoup4
certifi
charset-normalizer
colorama
contourpy
cycler
Cython~=3.0.11
decorator
exceptiongroup
fire
fonttools
h11
httpcore
httpx
idna
imageio
imgaug
importlib_resources
kiwisolver
lazy_loader
lmdb
lxml
matplotlib
networkx
numpy~=1.24.4
opencv-contrib-python
opencv-python~=4.10.0.84
opencv-python-headless
opt-einsum
packaging~=24.1
paddleocr~=2.8.1
paddlepaddle
pillow~=10.4.0
protobuf
pyclipper
pyparsing
python-dateutil
python-docx
PyWavelets
PyYAML
rapidfuzz
requests
scikit-image
scipy
shapely
six
sniffio
soupsieve
termcolor
tifffile
tqdm
typing_extensions
urllib3
zipp~=3.19.2

pip~=24.2
wheel~=0.44.0
setuptools~=72.1.0
platformdirs~=4.2.2
docx~=0.2.4